+++ test/dataProvider.ts
@@ -34,6 +34,8 @@ export const dataProvider = [
     { snippet: 'snippet34.php', unused: 0 },
     { snippet: 'snippet35.php', unused: 1 },
     { snippet: 'snippet36.php', unused: 0 },
+    { snippet: 'snippet37.php', unused: 0 },
+    { snippet: 'snippet38.php', unused: 0 },
 ]
 
 export const invalidPhpSyntaxDataProvider = [